% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfhazard-package.R
\docType{package}
\name{sbfhazard-package}
\alias{sbfhazard}
\alias{sbfhazard-package}
\title{sbfhazard}
\description{
Smooth backfitting estimators for additive and multiplicative hazard models,
including SBF and binning-based fitting paths.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LiHaozhe-rgb/sbfhazard}
  \item Report bugs at \url{https://github.com/LiHaozhe-rgb/sbfhazard/issues}
}

}
\author{
\strong{Maintainer}: Haozhe Li \email{236228545+LiHaozhe-rgb@users.noreply.github.com}

Authors:
\itemize{
  \item Haozhe Li \email{236228545+LiHaozhe-rgb@users.noreply.github.com}
}

}
\keyword{internal}
